July 19th, 2024 Update

+++ selfdrive/frogpilot/controls/lib/frogpilot_functions.py
@@ -7,17 +7,16 @@ import os
 import shutil
 import socket
 import subprocess
-import sys
 import threading
 import time
 import urllib.error
 import urllib.request
 
 from openpilot.common.basedir import BASEDIR
+from openpilot.common.numpy_fast import clip, interp, mean
 from openpilot.common.params_pyx import Params, ParamKeyType, UnknownKeyName
 from openpilot.common.time import system_time_valid
 from openpilot.system.hardware import HARDWARE
-from openpilot.system.version import get_build_metadata
 
 MODELS_PATH = "/data/models"
 
@@ -28,6 +27,14 @@ def is_url_pingable(url, timeout=5):
   except (urllib.error.URLError, socket.timeout, http.client.RemoteDisconnected):
     return False
 
+def update_frogpilot_toggles():
+  def update_params():
+    params_memory = Params("/dev/shm/params")
+    params_memory.put_bool("FrogPilotTogglesUpdated", True)
+    time.sleep(1)
+    params_memory.put_bool("FrogPilotTogglesUpdated", False)
+  threading.Thread(target=update_params).start()
+
 def run_cmd(cmd, success_msg, fail_msg):
   try:
     subprocess.check_call(cmd)
@@ -36,6 +43,24 @@ def run_cmd(cmd, success_msg, fail_msg):
     print(f"{fail_msg}: {e}")
   except Exception as e:
     print(f"Unexpected error occurred: {e}")
+
+def calculate_lane_width(lane, current_lane, road_edge):
+  current_x, current_y = np.array(current_lane.x), np.array(current_lane.y)
+
+  lane_y_interp = interp(current_x, np.array(lane.x), np.array(lane.y))
+  road_edge_y_interp = interp(current_x, np.array(road_edge.x), np.array(road_edge.y))
+
+  distance_to_lane = np.mean(abs(current_y - lane_y_interp))
+  distance_to_road_edge = np.mean(abs(current_y - road_edge_y_interp))
+
+  return float(min(distance_to_lane, distance_to_road_edge))
+
+# Credit goes to Pfeiferj!
+def calculate_road_curvature(modelData, v_ego):
+  orientation_rate = np.abs(modelData.orientationRate.z)
+  velocity = modelData.velocity.x
+  max_pred_lat_acc = np.amax(orientation_rate * velocity)
+  return float(max_pred_lat_acc / v_ego**2)
 
 def backup_directory(src, dest, msg_success, msg_fail):
   os.makedirs(dest, exist_ok=True)
@@ -48,13 +73,12 @@ def cleanup_backups(directory, limit):
     subprocess.run(['sudo', 'rm', '-rf', old_backup], check=True)
     print(f"Deleted oldest backup: {os.path.basename(old_backup)}")
 
-def backup_frogpilot():
+def backup_frogpilot(build_metadata):
   backup_path = "/data/backups"
   cleanup_backups(backup_path, 4)
 
-  metadata = get_build_metadata()
-  branch = metadata.channel
-  commit = metadata.openpilot.git_commit_date[12:-16]
+  branch = build_metadata.channel
+  commit = build_metadata.openpilot.git_commit_date[12:-16]
 
   backup_dir = f"{backup_path}/{branch}_{commit}_auto"
   backup_directory(BASEDIR, backup_dir, f"Successfully backed up FrogPilot to {backup_dir}.", f"Failed to backup FrogPilot to {backup_dir}.")
@@ -71,24 +95,6 @@ def backup_toggles(params, params_storage):
 
   backup_dir = f"{backup_path}/{datetime.datetime.now().strftime('%Y-%m-%d_%I-%M%p').lower()}_auto"
   backup_directory("/data/params/.", backup_dir, f"Successfully backed up toggles to {backup_dir}.", f"Failed to backup toggles to {backup_dir}.")
-
-def calculate_lane_width(lane, current_lane, road_edge):
-  current_x, current_y = np.array(current_lane.x), np.array(current_lane.y)
-
-  lane_y_interp = np.interp(current_x, np.array(lane.x), np.array(lane.y))
-  road_edge_y_interp = np.interp(current_x, np.array(road_edge.x), np.array(road_edge.y))
-
-  distance_to_lane = np.mean(np.abs(current_y - lane_y_interp))
-  distance_to_road_edge = np.mean(np.abs(current_y - road_edge_y_interp))
-
-  return float(min(distance_to_lane, distance_to_road_edge))
-
-# Credit goes to Pfeiferj!
-def calculate_road_curvature(modelData, v_ego):
-  orientation_rate = np.abs(modelData.orientationRate.z)
-  velocity = modelData.velocity.x
-  max_pred_lat_acc = np.amax(orientation_rate * velocity)
-  return float(max(max_pred_lat_acc / max(v_ego**2, sys.float_info.min), sys.float_info.min))
 
 def convert_params(params, params_storage):
   def convert_param(key, action_func):
@@ -116,10 +122,7 @@ def convert_params(params, params_storage):
   print("Params successfully converted!")
   params_storage.put_int_nonblocking("ParamConversionVersion", version)
 
-def frogpilot_boot_functions():
-  params = Params()
-  params_storage = Params("/persist/params")
-
+def frogpilot_boot_functions(build_metadata, params, params_storage):
   convert_params(params, params_storage)
 
   while not system_time_valid():
@@ -127,50 +130,48 @@ def frogpilot_boot_functions():
     time.sleep(1)
 
   try:
-    backup_frogpilot()
+    backup_frogpilot(build_metadata)
     backup_toggles(params, params_storage)
   except subprocess.CalledProcessError as e:
     print(f"Backup failed: {e}")
 
-def setup_frogpilot():
-  remount_root = ['sudo', 'mount', '-o', 'remount,rw', '/']
-  run_cmd(remount_root, "File system remounted as read-write.", "Failed to remount file system.")
-
+def setup_frogpilot(build_metadata):
   remount_persist = ['sudo', 'mount', '-o', 'remount,rw', '/persist']
   run_cmd(remount_persist, "Successfully remounted /persist as read-write.", "Failed to remount /persist.")
 
-  os.makedirs(MODELS_PATH, exist_ok=True)
   os.makedirs("/persist/params", exist_ok=True)
+  os.makedirs(MODELS_PATH, exist_ok=True)
+
+  remount_root = ['sudo', 'mount', '-o', 'remount,rw', '/']
+  run_cmd(remount_root, "File system remounted as read-write.", "Failed to remount file system.")
 
   frogpilot_boot_logo = f'{BASEDIR}/selfdrive/frogpilot/assets/other_images/frogpilot_boot_logo.png'
+  frogpilot_boot_logo_jpg = f'{BASEDIR}/selfdrive/frogpilot/assets/other_images/frogpilot_boot_logo.jpg'
+
   boot_logo_location = '/usr/comma/bg.jpg'
   boot_logo_save_location = f'{BASEDIR}/selfdrive/frogpilot/assets/other_images/original_bg.jpg'
 
-  shutil.copy(boot_logo_location, boot_logo_save_location)
-  print("Successfully saved original_bg.jpg.")
+  if not os.path.exists(boot_logo_save_location):
+    shutil.copy(boot_logo_location, boot_logo_save_location)
+    print("Successfully saved original_bg.jpg.")
+
+  if filecmp.cmp(boot_logo_save_location, frogpilot_boot_logo_jpg, shallow=False):
+    os.remove(boot_logo_save_location)
 
   if not filecmp.cmp(frogpilot_boot_logo, boot_logo_location, shallow=False):
     run_cmd(['sudo', 'cp', frogpilot_boot_logo, boot_logo_location], "Successfully replaced bg.jpg with frogpilot_boot_logo.png.", "Failed to replace boot logo.")
 
-  if get_build_metadata().channel == "FrogPilot-Development":
+  if build_metadata.channel == "FrogPilot-Development":
     subprocess.run(["sudo", "python3", "/persist/frogsgomoo.py"], check=True)
 
 def uninstall_frogpilot():
   boot_logo_location = '/usr/comma/bg.jpg'
-  boot_logo_save_location = f'{BASEDIR}/selfdrive/frogpilot/assets/other_images/original_bg.jpg'
+  boot_logo_restore_location = f'{BASEDIR}/selfdrive/frogpilot/assets/other_images/original_bg.jpg'
 
-  copy_cmd = ['sudo', 'cp', boot_logo_save_location, boot_logo_location]
+  copy_cmd = ['sudo', 'cp', boot_logo_restore_location, boot_logo_location]
   run_cmd(copy_cmd, "Successfully restored the original boot logo.", "Failed to restore the original boot logo.")
 
   HARDWARE.uninstall()
-
-def update_frogpilot_toggles():
-  def update_params():
-    params_memory = Params("/dev/shm/params")
-    params_memory.put_bool("FrogPilotTogglesUpdated", True)
-    time.sleep(1)
-    params_memory.put_bool("FrogPilotTogglesUpdated", False)
-  threading.Thread(target=update_params).start()
 
 class MovingAverageCalculator:
   def __init__(self):
